implementation_plan.py
Here's an implementation approach using MCP for your Mixpanel AI agent:

```python
# Main application structure using MCP for a Mixpanel AI agent

import os
from typing import List, Dict, Any
from datetime import datetime, timedelta

# MCP core imports
from mcp import Agent, Context, Action, Tool, Memory
from mcp.tools import APITool, DataProcessingTool
from mcp.memory import ConversationBuffer

# Integration libraries
import mixpanel
from sendgrid import SendGridAPIClient
from sendgrid.helpers.mail import Mail

# Initialize Mixpanel client
mixpanel_client = mixpanel.Mixpanel(os.getenv("MIXPANEL_TOKEN"))

# Define Tools for the agent

class MixpanelDataTool(APITool):
    """Tool for fetching and analyzing Mixpanel data"""
    
    def get_struggling_users(self, feature_name: str, threshold: float = 0.3, days: int = 7) -> List[Dict]:
        """
        Identify users struggling with a specific feature based on engagement patterns
        """
        # Get event data for the feature from Mixpanel
        end_date = datetime.now()
        start_date = end_date - timedelta(days=days)
        
        # JQL query to identify struggling users based on:
        # 1. High number of attempts on feature
        # 2. Low completion rate
        # 3. Error events associated with feature
        query = f"""
        function main() {{
            return Events({{
                from_date: '{start_date.strftime("%Y-%m-%d")}',
                to_date: '{end_date.strftime("%Y-%m-%d")}',
                event_selectors: [{{event: "{feature_name}_attempt"}}, {{event: "{feature_name}_complete"}}, {{event: "{feature_name}_error"}}]
            }})
            .groupByUser(["properties.distinct_id", "properties.email", "properties.name"], {{
                attempts: mixpanel.reducer.count(event => event.name == "{feature_name}_attempt"),
                completions: mixpanel.reducer.count(event => event.name == "{feature_name}_complete"),
                errors: mixpanel.reducer.count(event => event.name == "{feature_name}_error")
            }})
            .filter(user => user.attempts > 3 && (user.completions / user.attempts) < {threshold})
            .map(user => ({{
                user_id: user.key[0],
                email: user.key[1], 
                name: user.key[2],
                attempts: user.value.attempts,
                completions: user.value.completions,
                completion_rate: user.value.completions / user.value.attempts,
                errors: user.value.errors
            }}));
        }}
        """
        
        results = mixpanel_client.query("jql", {"script": query})
        return results

class ContentGenerationTool(Tool):
    """Tool for generating personalized tutorial content"""
    
    def create_tutorial(self, feature_name: str, user_data: Dict, struggle_metrics: Dict) -> Dict:
        """
        Generate personalized tutorial content based on user struggles
        """
        # Context building for the MCP agent
        context = Context()
        
        # Add user context
        context.add("user_name", user_data.get("name", "there"))
        context.add("feature_name", feature_name)
        context.add("struggle_points", {
            "attempts": struggle_metrics.get("attempts", 0),
            "completion_rate": struggle_metrics.get("completion_rate", 0),
            "errors": struggle_metrics.get("errors", 0)
        })
        
        # Fetch product documentation about the feature
        # (Simplified for MVP - would connect to knowledge base)
        feature_docs = self._get_feature_documentation(feature_name)
        context.add("feature_documentation", feature_docs)
        
        # Use MCP to generate personalized content
        prompt = """
        Create a personalized tutorial to help a user who is struggling with a product feature.
        The tutorial should be friendly, helpful, and specifically address their struggle points.
        Include clear step-by-step instructions with examples.
        """
        
        result = context.generate(prompt)
        
        return {
            "subject": f"Quick help with {feature_name}",
            "content": result.text,
            "format": "html"  # Could be markdown, plain, etc.
        }
    
    def _get_feature_documentation(self, feature_name: str) -> str:
        # In a real implementation, this would fetch docs from a knowledge base
        # Simplified for MVP
        return f"Documentation for {feature_name}..."

class CommunicationTool(Tool):
    """Tool for sending communications to users"""
    
    def __init__(self):
        self.sg_client = SendGridAPIClient(os.getenv("SENDGRID_API_KEY"))
    
    def send_email(self, recipient: Dict, content: Dict) -> bool:
        """Send an email to a user with tutorial content"""
        message = Mail(
            from_email=os.getenv("FROM_EMAIL", "help@yourproduct.com"),
            to_emails=recipient["email"],
            subject=content["subject"],
            html_content=content["content"]
        )
        
        try:
            response = self.sg_client.send(message)
            return response.status_code >= 200 and response.status_code < 300
        except Exception as e:
            print(f"Error sending email: {e}")
            return False

# Main Agent Class

class MixpanelAssistantAgent(Agent):
    """Proactive AI agent that helps users struggling with product features"""
    
    def __init__(self):
        super().__init__(
            tools=[
                MixpanelDataTool(),
                ContentGenerationTool(),
                CommunicationTool()
            ],
            memory=ConversationBuffer(max_tokens=2000)
        )
    
    def run_daily_check(self, features_to_monitor: List[str]):
        """Check for struggling users across monitored features and assist them"""
        
        for feature in features_to_monitor:
            # Create a new context for this run
            context = Context()
            context.add("feature", feature)
            context.add("check_time", datetime.now().isoformat())
            
            # Step 1: Identify struggling users
            mixpanel_tool = self.get_tool("MixpanelDataTool")
            struggling_users = mixpanel_tool.get_struggling_users(feature)
            
            if not struggling_users:
                print(f"No users struggling with {feature}")
                continue
            
            # Step 2: For each struggling user, generate and send help
            content_tool = self.get_tool("ContentGenerationTool")
            comms_tool = self.get_tool("CommunicationTool")
            
            for user in struggling_users:
                # Generate personalized tutorial
                tutorial = content_tool.create_tutorial(
                    feature_name=feature,
                    user_data={"name": user.get("name"), "email": user.get("email")},
                    struggle_metrics={
                        "attempts": user.get("attempts"),
                        "completion_rate": user.get("completion_rate"),
                        "errors": user.get("errors")
                    }
                )
                
                # Send the tutorial
                sent = comms_tool.send_email(
                    recipient={"email": user.get("email")},
                    content=tutorial
                )
                
                # Log the action and result
                self.memory.add(
                    Action(
                        name="assist_user",
                        params={
                            "user_id": user.get("user_id"),
                            "feature": feature,
                            "tutorial_sent": sent,
                            "timestamp": datetime.now().isoformat()
                        }
                    )
                )
                
                print(f"Assisted user {user.get('user_id')} with {feature}: Success={sent}")

# Usage example

if __name__ == "__main__":
    # Setup agent
    agent = MixpanelAssistantAgent()
    
    # Run daily check for features we want to monitor
    agent.run_daily_check([
        "data_export", 
        "custom_report", 
        "user_cohorts"
    ])

```

This implementation demonstrates how to build your Mixpanel AI agent using MCP. Here's a breakdown of the key components:

1. **Core Architecture**:
   - Uses MCP's `Agent`, `Context`, `Tool`, and `Memory` abstractions
   - Implements three essential tools: data retrieval, content generation, and communication

2. **Mixpanel Integration**:
   - The `MixpanelDataTool` uses JQL (JSON Query Language) to identify struggling users
   - Looks for patterns like high attempt counts, low completion rates, and errors

3. **Content Generation**:
   - `ContentGenerationTool` uses MCP's context management to build prompts
   - Incorporates user-specific struggle metrics and product documentation

4. **Communication**:
   - `CommunicationTool` handles email delivery through SendGrid
   - Could be expanded to support other channels as needed

5. **Workflow Logic**:
   - The `run_daily_check` method orchestrates the entire process
   - Processes each monitored feature, identifies struggling users, and provides assistance
   - Maintains memory of actions taken for future reference

For your MVP, you would need to:
1. Set up the necessary environment variables (API keys, etc.)
2. Define the specific features you want to monitor
3. Customize the JQL queries based on your Mixpanel event structure
4. Enhance the content generation with more specific product knowledge

This implementation provides a solid foundation that could be deployed as a scheduled task (via cron, AWS Lambda, etc.) to proactively help users on a daily basis. It uses MCP's context management and tool integration to make the code cleaner and more maintainable than building everything from scratch.

Would you like me to elaborate on any specific part of this implementation?